+++ orchestra/src/skills/index.ts
@@ -1,4 +1,3 @@
-export * from "./builtin";
 export * from "./convert";
 export * from "./crud";
 export * from "./events";